ingest/requirements.txt
# Ingestion pipeline
pymupdf4llm>=0.0.17
langchain-text-splitters>=0.3.0
google-genai>=0.8.0

# Elasticsearch
elasticsearch>=8.0.0,<9.0.0

# Queue
redis>=5.0.0
rq>=1.16.0

# Format extractors
python-docx>=1.1.0       # .docx
openpyxl>=3.1.0          # .xlsx
beautifulsoup4>=4.12.0   # .html
lxml>=5.0.0              # faster HTML parser for bs4 (optional but recommended)
